+++ Extensions/PlayerExtensions.cs
@@ -1,5 +1,7 @@
 ﻿using Bark.Tools;
 using GorillaLocomotion;
+using HarmonyLib;
+using Photon.Pun;
 using UnityEngine;
 
 namespace Bark.Extensions
@@ -14,6 +16,11 @@ namespace Bark.Extensions
         public static void SetVelocity(this Player self, Vector3 v)
         {
             self.bodyCollider.attachedRigidbody.velocity = v;
+        }
+
+        public static PhotonView PhotonView(this VRRig rig)
+        {
+            return Traverse.Create(rig).Field("photonView").GetValue<PhotonView>();
         }
     }
 }