+++ Modules/Multiplayer/X-Ray.cs
@@ -1,6 +1,7 @@
 ﻿using Bark.Extensions;
 using Bark.GUI;
 using Bark.Tools;
+using System;
 using System.Collections.Generic;
 using UnityEngine;
 
@@ -15,13 +16,24 @@ namespace Bark.Modules.Multiplayer
         {
             foreach (var rig in GorillaParent.instance.vrrigs)
             {
-                if (rig.photonView.Owner.IsLocal) continue;
+                try
+                {
+                    if (rig.PhotonView().Owner.IsLocal) continue;
 
-                var marker = rig.gameObject.GetComponent<XRayMarker>();
-                if (marker)
-                    marker.Update();
-                else
-                    markers.Add(rig.gameObject.AddComponent<XRayMarker>());
+                    var marker = rig.gameObject.GetComponent<XRayMarker>();
+                    if (marker)
+                        marker.Update();
+                    else
+                        markers.Add(rig.gameObject.AddComponent<XRayMarker>());
+                }
+                catch (Exception e)
+                {
+                    Logging.LogException(e);
+                    Logging.LogDebug("rig is null:", rig is null);
+                    Logging.LogDebug("rig?.PhotonView() is null:", rig?.PhotonView() is null);
+                    Logging.LogDebug("rig?.PhotonView()?.Owner is null:", rig?.PhotonView()?.Owner is null);
+                    Logging.LogDebug("rig?.gameObject is null:", rig?.gameObject is null);
+                }
             }
         }