+++ Patches/SpeedPatches.cs
@@ -3,6 +3,9 @@ using Bark.Modules.Movement;
 using Bark.Modules;
 using Bark.Tools;
 using System;
+using AA;
+using GorillaLocomotion;
+using UnityEngine;
 
 namespace Bark.Patches
 {
@@ -69,6 +72,21 @@ namespace Bark.Patches
 
                 for (int i = 0; i < __result.Length; i++)
                     __result[i] *= SpeedBoost.scale;
+            }
+            catch (Exception e) { Logging.LogException(e); }
+        }
+    }
+
+    [HarmonyPatch(typeof(Player))]
+    [HarmonyPatch("GetSwimmingVelocityForHand", MethodType.Normal)]
+    internal class SwimmingVelocityPatch
+    {
+        private static void Postfix(ref Vector3 swimmingVelocityChange)
+        {
+            try
+            {
+                if (!SpeedBoost.active) return;
+                swimmingVelocityChange *= SpeedBoost.scale;
             }
             catch (Exception e) { Logging.LogException(e); }
         }